finished premium places functionality endpoint

+++ SkimSpotsApi/app/controllers/premiumPlace.controller.js
@@ -35,41 +35,24 @@ exports.getAll = (req, res) => {
 exports.getCurrentPremiumPlaces = (req, res) => {
     PremiumPlaces.findAll()
         .then((data) => {
-            // TODO: consider empty premium places array
-            const premiumPlacesDate = data[0].date.toISOString().slice(0, 10);
-            const todayDate = new Date().toISOString().slice(0, 10);
-
-            if (premiumPlacesDate === todayDate) {
-                res.send(data);
-            } else {
-                // Clear all items in premium places table
-                PremiumPlaces.destroy({
-                    truncate: true,
-                }).then(() => {
-                    // Get all places
-                    Places.findAll().then((data) => {
-                        const allPlaces = data;
-                        // Get randomly 2 places
-                        const randomlySelectedPlaces = getRandomItemsFromArray(
-                            allPlaces,
-                            2
-                        );
-                        const newPremiumPlaces = [];
-
-                        // Map selected random places and add it to new premium places array
-                        randomlySelectedPlaces.map((placeId) => {
-                            newPremiumPlaces.push({
-                                date: new Date(),
-                                premiumPoints: 20,
-                                placeId: placeId,
-                            });
-                        });
-
-                        // Add multiple(2) items to empty premium places table
-                        PremiumPlaces.bulkCreate(newPremiumPlaces).then((data) => {
+            if (data.length > 0) {
+                const premiumPlacesDate = data[0].date.toISOString().slice(0, 10);
+                const todayDate = new Date().toISOString().slice(0, 10);
+                if (premiumPlacesDate === todayDate) {
+                    res.send(data);
+                } else {
+                    // Clear all items in premium places table
+                    PremiumPlaces.destroy({
+                        truncate: true,
+                    }).then(() => {
+                        createNewPremiumPlaces().then((data) => {
                             res.send(data);
                         });
                     });
+                }
+            } else {
+                createNewPremiumPlaces().then((data) => {
+                    res.send(data);
                 });
             }
         })
@@ -78,6 +61,34 @@ exports.getCurrentPremiumPlaces = (req, res) => {
                 message: err.message || "Some error occurred while retrieving UserPlaces.",
             });
         });
+};
+
+const createNewPremiumPlaces = () => {
+    return new Promise((resolve) => {
+        // Get all places
+        Places.findAll().then((data) => {
+            const allPlaces = data;
+            // Get randomly 2 places
+            const randomlySelectedPlaces = getRandomItemsFromArray(allPlaces, 2);
+            const newPremiumPlaces = [];
+
+            // Map selected random places and add it to new premium places array
+            const premiumPoints = [30, 15];
+
+            randomlySelectedPlaces.map((placeId, index) => {
+                newPremiumPlaces.push({
+                    date: new Date(),
+                    premiumPoints: premiumPoints.length >= index + 1 ? premiumPoints[index] : 10,
+                    placeId: placeId,
+                });
+            });
+
+            // Add multiple(2) items to empty premium places table
+            PremiumPlaces.bulkCreate(newPremiumPlaces).then((data) => {
+                resolve(data);
+            });
+        });
+    });
 };
 
 const getRandomItemsFromArray = (array, itemsCount) => {
